+++ HISTORY.md
@@ -128,3 +128,6 @@
 - v0.0.0:47
 - main.go에서 answer를 보내도록 개선
 - v0.0.0:48
+- offer, answer 전달
+    -  보완할 점 : signalServer에서 접속하는 유저들에게 저장해놓은 offer를 매번 쏴주자. 
+- v0.0.0:49